requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
sumy==0.11.0
arxiv==1.4.4
scholarly==1.7.11
google-api-python-client==2.100.0
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
lxml==4.9.3
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
arxiv==2.1.0
gunicorn==21.2.0
python-dotenv==1.0.0